pyapi/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
spacy==3.8.7
openai==2.0.0
python-dotenv==1.1.1